remove yetAnotherUselessCast
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@16308 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/keys/InsertableClientSSK.java
@@ -134,7 +134,7 @@ public class InsertableClientSSK extends ClientSSK {
         headers[x++] = (byte) (KeyBlock.HASH_SHA256);
         // Then crypto ID
         headers[x++] = (byte) (Key.ALGO_AES_PCFB_256_SHA256 >> 8);
-        headers[x++] = (byte) (Key.ALGO_AES_PCFB_256_SHA256);
+        headers[x++] = Key.ALGO_AES_PCFB_256_SHA256;
         // Then E(H(docname))
 		// Copy to headers
 		System.arraycopy(ehDocname, 0, headers, x, ehDocname.length);
